requirements.txt
vkbottle
pillow
lxml
beautifulsoup4
requests
peewee